+++ src/testmod/java/me/basiqueevangelist/dynreg/testmod/desc/FlowerPotBlockEntry.java
@@ -1,7 +1,9 @@
 package me.basiqueevangelist.dynreg.testmod.desc;
 
 import com.google.gson.JsonObject;
-import me.basiqueevangelist.dynreg.entry.*;
+import me.basiqueevangelist.dynreg.api.entry.EntryRegisterContext;
+import me.basiqueevangelist.dynreg.api.entry.EntryScanContext;
+import me.basiqueevangelist.dynreg.api.entry.RegistrationEntry;
 import me.basiqueevangelist.dynreg.wrapped.SimpleHashers;
 import me.basiqueevangelist.dynreg.wrapped.SimpleReaders;
 import me.basiqueevangelist.dynreg.testmod.DynRegTest;
@@ -57,8 +59,8 @@ public class FlowerPotBlockEntry implements RegistrationEntry {
     }
 
     @Override
-    public int hash() {
-        int hash = id.hashCode();
+    public long hash() {
+        long hash = id.hashCode();
         hash = 31 * hash + content.hashCode();
         hash = 31 * hash + SimpleHashers.hash(settings);
         return hash;
